# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "chat-API",
    "description": "...",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/users/login": {
      "post": {
        "tags": ["users"],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "ian.rosasmal@gmail.com",
                      "password": "Academlo123@"
                    }
                  }
                ]
              },
              "example": {
                "email": "ian.rosasmal@gmail.com",
                "password": "Academlo123@"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/messages/conversation/{id}": {
      "post": {
        "tags": ["messages"],
        "summary": "create new message",
        "operationId": "createnewmessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createnewmessagerequest"
                  },
                  {
                    "example": {
                      "content": "que te importa",
                      "senderId": 6
                    }
                  }
                ]
              },
              "example": {
                "content": "que te importa",
                "senderId": 6
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "tags": ["messages"],
        "summary": "get conversation messages",
        "operationId": "getconversationmessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "ian.rosasmal@gmail.com",
          "password": "Academlo123@"
        }
      },
...
      "createnewmessagerequest": {
        "title": "createnewmessagerequest",
        "required": ["content", "senderId"],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "senderId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "content": "que te importa",
          "senderId": 6
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}